fix: created and separated a new layout for the header
