+++ src/common/header/components/loginOption/style.ts
@@ -1,0 +1,7 @@
+import styled from '@emotion/styled';
+
+export const LoginOption = styled.div`
+  display: flex;
+  flex-direction: row;
+  gap: 15px;
+`;
