+++ src/common/header/index.tsx
@@ -1,16 +1,12 @@
 import { Outlet } from 'react-router-dom';
 import * as s from './style';
 import { Right } from './right';
+import { Left } from './left';
 export const Header = () => {
   return (
     <>
       <s.Header>
-        <s.LeftSide>
-          <a href="/" className="mainTitle">
-            GameCrew
-          </a>
-          <span>Let's play an exciting game</span>
-        </s.LeftSide>
+        <Left />
         <Right />
       </s.Header>
       <Outlet />
