+++ src/common/header/left/index.tsx
@@ -1,0 +1,11 @@
+import * as s from './style';
+export const Left = () => {
+  return (
+    <s.Left>
+      <a href="/" className="mainTitle">
+        GameCrew
+      </a>
+      <span>Let's play an exciting game</span>
+    </s.Left>
+  );
+};
