+++ src/common/header/right/style.ts
@@ -1,7 +1,7 @@
 import { Theme } from '@/styles/theme';
 import styled from '@emotion/styled';
 
-export const RightSide = styled.div`
+export const Right = styled.div`
   height: 100%;
   min-width: 600px;
   width: fit-content;
@@ -22,21 +22,4 @@ export const LoginOption = styled.div`
   display: flex;
   flex-direction: row;
   gap: 15px;
-`;
-export const UserData = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 15px;
-
-  .nickname {
-    font-size: ${Theme.fontSizes.body2};
-    font-weight: ${Theme.fontWeights.bold};
-  }
-
-  .temp {
-    font-size: ${Theme.fontSizes.body4};
-    font-weight: ${Theme.fontWeights.normal};
-    color: ${Theme.colors.gray2};
-  }
 `;
