+++ src/common/header/style.ts
@@ -17,26 +17,3 @@ export const Header = styled.div`
   color: ${Theme.colors.white};
   box-shadow: 10px 10px 10px ${Theme.colors.gray2};
 `;
-
-export const LeftSide = styled.div`
-  background-color: ${Theme.colors.gray3};
-  min-width: 400px;
-  height: 100%;
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  gap: 20px;
-  .mainTitle {
-    font-size: ${Theme.fontSizes.h1};
-    font-weight: ${Theme.fontWeights.bold};
-    color: ${Theme.colors.white};
-    text-decoration: none;
-  }
-  & span {
-    font-size: ${Theme.fontSizes.body2};
-    font-weight: ${Theme.fontWeights.normal};
-  }
-`;
